% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support-fxns.R
\name{get_wind_profiles}
\alias{get_wind_profiles}
\title{Return wind profile data from HURDAT2 observation}
\usage{
get_wind_profiles(obs)
}
\arguments{
\item{obs}{SpatialPointsDataFrame: a single hurricane observation represented by
a single row of an object returned from \code{load_hurdat_track()}}
}
\description{
This function takes a single HURDAT2 observation retreived from \code{load_hurdat_track()}
and returns a wind profileis used to download, parse, and format the current tropical storm database
}
\examples{
# Download recent data from HURDAT2 (NOAA)
fetch_best_tracks_data(hurdat_database)
track = load_hurdat_track('AL142018') # Load data from hurricane Micahel
observation = track[17,] #select a single observation to analyze
get_wind_profiles(observation)
}
