% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurrecon.R
\name{hurrecon_run}
\alias{hurrecon_run}
\title{Run HURRECON model for single Hurricane Track}
\usage{
hurrecon_run(
  trk,
  max_rad_km = 100,
  res_m = 500,
  max_interp_dist_km = 1,
  proj = "32616",
  aoi = NULL
)
}
\arguments{
\item{trk}{SpatialPointsDataFrame of single track data returned from \code{load_hurdat_track}}

\item{max_rad_km}{integer: maximum distance away from eye to model windspeed}

\item{res_m}{integer: spatial resolution of output raster}

\item{max_interp_dist_km}{integer: with more than one observation, \code{max_interp_dist_km}
is the spatial resolution of the linear interpolation between concurrent hurrecon observation.
When \code{max_interp_dist_km} = 1, an interpolated observation is placed every 1 km between two
observations further apart than 1 km. Values < 1 km are not recommend to avoid gaps between
wind fields}

\item{proj}{string: default projecgtion for analyses. Recommend keeping at UTM16 (epsg:32616)}

\item{aoi}{area of interest, track locations outside of aoi are ignored. This can be used
to limit results for a specific spatial region (county, state, country, etc.). Must be
in same projection as \code{trk}.}

\item{land}{SpatialPolygons: This is spatial layer representing land. This is used to identify when the hurricane is on land to incorporate appropriate adjustment factors}
}
\description{
This function takes a single hurricane track from \code{load_hurdat_track} and
estimates a wind field producing the maximum sustained wind (\code{Vs} in ms^-1)
and direction of maximum wind speed (\code{d} in radians)
}
\examples{
# Download recent data from HURDAT2 (NOAA)
library(terra)
library(hurrecon)
data("geographic")

path = 'hurdat_data.csv'
fetch_best_tracks_data(path)

# load data for trackID AL142018 (Hurricane Michael)
track = load_hurdat_track(path, trackID = 'AL142018')
 
# Quick one for example
output = hurrecon_run(track, land=land, max_rad_km = 100, res_m = 500, max_interp_dist_km = 50)
plot(land)
plot(output, add = TRUE)
plot(land)

}
