% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-parse-hurdat-data.R
\name{load_hurdat_track}
\alias{load_hurdat_track}
\title{Load Full Track and Return as Spatial Points Dataframe}
\usage{
load_hurdat_track(path, trackID, proj = 32616)
}
\arguments{
\item{path}{character: path to parsed HURDAT2 database downloaded using \code{fetch_best_tracks_data()}.}

\item{trackID}{string: trackID from HURDAT database (e.g., AL122018 indicates Hurricane Michael, the twelfth Atlantic hurricane of the 2018 season)}

\item{proj}{output CRS for Spatial Points Dataframe. Must be a projection coordinate system (e.g., UTM) for distance-based calculations}
}
\description{
This function returns an attributed shapefile containing tropical cyclone positions
and attributes from HURDAT2 database.
}
\examples{
#' # Download recent data from HURDAT2 (NOAA)
path = 'hurdat_data.csv'
fetch_best_tracks_data(path)

# load data for trackID AL142018 (Hurricane Michael)
track = load_hurdat_track(path, trackID = 'AL142018')

# Retrieve a US shapefile and project to UTM16
library(maptools)
data("wrld_simpl")
us = wrld_simpl['USA',]
us = spTransform(us, '+init=epsg:32616')

output_raster = hurrecon_run(track, land=us, max_rad_km = 100, res_m = 500, max_interp_dist_km = 1)
raster::plot(output_raster)

}
