% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-parse-hurdat-data.R
\name{load_hurdat_track}
\alias{load_hurdat_track}
\title{Load Full Track and Return as POINT simple feature}
\usage{
load_hurdat_track(path, trackID, proj = 32616)
}
\arguments{
\item{path}{character: path to parsed HURDAT2 database downloaded using \code{fetch_best_tracks_data()}.}

\item{trackID}{string: trackID from HURDAT database (e.g., AL122018 indicates Hurricane Michael, the twelfth Atlantic hurricane of the 2018 season)}

\item{proj}{output CRS for feature. Must be a projection coordinate system (e.g., UTM) for distance-based calculations}
}
\description{
This function returns an attributed feature containing tropical cyclone positions
and attributes from HURDAT2 database.
}
\examples{
# Download recent data from HURDAT2 (NOAA)
library(terra)
library(hurrecon)
data("geographic")

path = 'hurdat_data.csv'
fetch_best_tracks_data(path)

# load data for trackID AL142018 (Hurricane Michael)
track = load_hurdat_track(path, trackID = 'AL142018')
 
# Quick one for example
output = hurrecon_run(track, land=land, max_rad_km = 100, res_m = 500, max_interp_dist_km = 50)
plot(land)
plot(output, add = TRUE)
plot(land)

}
